Move metric functions to a separate file
